Style refactor
