+++ platform/ui/src/components/userPreferences/GeneralPreferences.styl
@@ -1,0 +1,18 @@
+.GeneralPreferences
+  display: flex;
+  padding: 20px 0;
+
+  .language
+    display: flex;
+    width: 50%;
+
+    .languageLabel
+      display: flex;
+      flex-basis: 0;
+      flex-grow: 1.5;
+      justify-content: flex-end;
+      padding-right: 20px;
+
+    .language-select
+      flex-grow: 1.5;
+      flex-basis: 0;
